services/window_collector/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
kafka-python==2.0.2
pandas==2.2.2
pyarrow==15.0.2
influxdb-client